+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bump runtimeVersion to `2025.05.21`.
  * Upgraded runtime Dart SDK to `3.8.0`.
  * Upgraded stable Flutter analysis SDK to `3.32.0`.
+ * Note: started to backfill `Package[Version]`'s `isAdminDeleted` field.
 
 ## `20250519t093200-all`